+++ Gruntfile.js
@@ -14,12 +14,28 @@ module.exports = function (grunt) {
         files : [{
           expand  : true,
           cwd     : 'src/',
-          src     : '**/*.js',
-          dest    : 'dist/'
+          src     : [ '**/*.js' ],
+          dest    : 'dist/',
+          exceptionsFiles : [ 'converter.json' ]
         }]
       }
     },
-
+    /**
+     * Copy needed file to correct dist directory
+     */
+    copy      : {
+      json  : {
+        files : [
+          // makes all src relative to cwd
+          {
+            expand  : true,
+            cwd     : 'src/',
+            src     : [ '**/*.json' ],
+            dest    : 'dist/'
+          }
+        ]
+      }
+    },
     /**
      * Mocah unit test
      */
@@ -40,12 +56,13 @@ module.exports = function (grunt) {
   });
 
   grunt.loadNpmTasks('grunt-contrib-uglify');
+  grunt.loadNpmTasks('grunt-contrib-copy');
   grunt.loadNpmTasks('grunt-mocha-cli');
-  grunt.loadNpmTasks('yocto-hint');
+  //grunt.loadNpmTasks('yocto-hint');
 
   // register tasks
   grunt.registerTask('hint', 'yoctohint');
   grunt.registerTask('test', 'mochacli');
-  grunt.registerTask('build', [ 'hint', 'uglify' ]);
+  grunt.registerTask('build', [ 'uglify', 'copy' ]);
   grunt.registerTask('default', [ 'build', 'test' ]);
 };