+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0 (2017-07-25)
+
+- Rewrite all source code to use a new workflow
+- Update nodemailer to latest version
+- Update mandrill API to latest version
+
 ## 1.2.4 (2017-01-18)
 -  Fix module version in package.json because new version of module was not compatible with node v4.x
   - "striptags": "2.1.1"